+++ src/present.rs
@@ -10,7 +10,7 @@ use winit::event::WindowEvent;
 use winit::event_loop::{ActiveEventLoop, ControlFlow, EventLoop};
 use winit::window::{Window, WindowId};
 
-use crate::utils;
+use crate::{utils, Float};
 use crate::vec3::Vec3;
 
 #[derive(Debug)]
@@ -25,11 +25,11 @@ pub struct Presentation {
     surface: Option<Surface<Rc<Window>, Rc<Window>>>,
     width: u32,
     height: u32,
-    samples: f64,
+    samples: Float,
 }
 
 impl Presentation {
-    pub fn new(width: u32, height: u32, samples: f64) -> Self {
+    pub fn new(width: u32, height: u32, samples: Float) -> Self {
         println!("presentation width {} height {} ", width, height);
         Presentation {
             window: None,
@@ -102,9 +102,9 @@ impl ApplicationHandler<PresentationEvent> for Presentation {
         let mut b = event.color.z;
 
         let scale = 1.0 / self.samples;
-        r = f64::sqrt(scale * r);
-        g = f64::sqrt(scale * g);
-        b = f64::sqrt(scale * b);
+        r = Float::sqrt(scale * r);
+        g = Float::sqrt(scale * g);
+        b = Float::sqrt(scale * b);
         let r = (256.0 * r.clamp(0.0, 0.999)) as u32;
         let g = (256.0 * g.clamp(0.0, 0.999)) as u32;
         let b = (256.0 * b.clamp(0.0, 0.999)) as u32;
